fix unresponsive video container (#1091)
Summary:
There was a video container on the documentation site that was not responsive in smaller devices and was not looking good.

Currently it looks something like this on smaller screen.
![image](https://github.com/facebook/hermes/assets/99970842/54ec540f-3687-45dc-983e-962c2bdcc306)

So, I have fixed that unresponsive video container and after the fix, it looks like this.

![image](https://github.com/facebook/hermes/assets/99970842/86f84c95-bb34-404a-9a75-c8d44ef4761a)

This is my first pull request on open source. So, if I have missed something please let me know.

Pull Request resolved: https://github.com/facebook/hermes/pull/1091

Reviewed By: yungsters

Differential Revision: D48479906

Pulled By: mattbfb

fbshipit-source-id: 0af8ba1d5d8bcf91174f10a02e80396cbfe12402