Fix TypeOperator and Keyof parsing precedence
Summary:
D48700489 noted that the precedence of the original implementation
would interpret `keyof number | string` as `keyof (number | string)`
instead of `(keyof number) | string`. The fix used
`parsePrimaryFlowTypeAnnotation` to reduce the precedence.

But primary  type annotations are a little too low-- this disallows
`?Maybe` types  and `Generic.QualifiedTypes`.
This diff instead uses the `parsePrefixFlowTypeAnnotation` method,
which will allow for both of those cases

Reviewed By: avp

Differential Revision: D48737810

fbshipit-source-id: e175b9cd68699fa50eae66c76b5606bfe6d753c3