Add NativeCallInst
Summary:
`NativeCallInst` implements a native call. Note that it is not a
subclass of `BaseCallInst`, since this is conceptually a very different
operation.

Its parameters are the callee, which must be a native function pointer,
a native function signature, so the code generator knows exactly how to
emit the call, and a list of JS values for the parameters. The JS values
will be converted to native during the call, and the native result will
be converted back to JS representation (these operations are cheap).

Reviewed By: avp

Differential Revision: D48720350

fbshipit-source-id: 502e82a744ab59050f89fb438417e533c263b449