Flow: add new types CPtr and NativeFunction
Summary:
`CPtr` represents an untyped native data pointer, essentially
`(void *)`.

`NativeFunction` is a sibling of `TypedFunction` and represents a
native function pointer to a native function, for example
`(int (*)(const char *))`. It has an associated native signature with
the exact native types of its parameters and result, but it also
has JS types similar to `TypedFunction`. So every native parameter and
result is mapped to a JS type: integers and floating point to `Number`,
pointers to `CPtr`, etc.

`NativeFunction` is not compatible with `TypedFunction`, i.e. one cannot
flow into the other.

Both `NativeFunction` and `CPtr` will actually be encoded as HermesValue
number at runtime, because the bit pattern is compatible, i.e. the bit
pattern of a valid pointer can never be interpreted as FP NaN (which we
will enforce at runtime).

Note that nothing is generating these types yet.

Reviewed By: neildhar

Differential Revision: D48720355

fbshipit-source-id: f720d43d733e71945a78a48e20a74e32e9273626