FlowChecker: add $SHBuiltin.extern_c()
Summary:
The `extern_c()` builtin declares a native external function with a
specific name and signature and returns a `NativeFunction` pointer with
the specified signature. This is the only way to construct a
`NativeFunction` value.

The overall syntax is:
```
$SHBuiltin.extern_c(options, function NAME(...): type {throw 0;});
```

## Options

`options` is an object literal and it provides a way to customize
the external declaration. For now the supported options deal with the
reality that we are emitting C instead of LLVM.

`declared: true|false`: this field tells the compiler that the external
function that we want to use is already declared in the C compilation
unit, most likely through a header. It would be an error to generate a
second `extern` declaration with types that do not match completely.

For example, let's imagine that the emitted C code already included
`<stdio.h>` (it doesn't, but this is a thought experiment). If we
declared an external function `fopen()`, Static Hermes would emit an
external declaration looking something like this:
```
extern void *fopen(void *, void *);
```
This would result in a Clang error, since this declaration conflicts
with the prior declaration of `fopen()` from `<stdio.h>`.

`include: string`: this field tells the compiler to emit an `#include`
directive with the specified string. The header will provide the extern
declaration.

This is necessary when importing built-in functions. Again, if we
generated an extern declaration for `fopen()` ourselves, Clang will emit
an error like:
```
warning: declaration of built-in function 'fopen' requires inclusion of
the header <stdio.h>
```

`include` implies `declared`.

If neither of these options are provided, Static Hermes will emit the
external declaration itself.

In the future we will add more options for calling conventions, function
attributes, etc.

## Function

The second argument of the builtin is a function expression. Its name
determines the name of the extern. Its parameters and return type
determine the precise native signature of the extern. They must all be
native types, except the return type, which may be `void`.

The body of the function expression is ignored, but it still undergoes
typechecking, so it must be valid. The convention is to always place a
`throw 0;` in the body to avoid worrying about the type of return.

## Semantics

The `extern_c` builtin declares an external function with the specified
name, signature and attributes and returns a *constant value* pointing
to the function. The same extern can be declared multiple times with no
ill effects, as long as the name and signatures match.

The result of the extern may even be called directly, for example like
this:
```
$SHBuiltin.extern_c({}, function foo(p: c_int): void {})(123);
```

The recommended convention is to assign all externs to module-scope
constants prefixed with underscore. For example, the recommended way to
import `fopen()` is like this:
```
const _fopen = $SHBuiltin.extern_c({include:'stdio.h'},
  function fopen(path: c_ptr, mode: c_ptr): c_ptr { throw 0; });
```

Reviewed By: avp

Differential Revision: D48720356

fbshipit-source-id: b674c757f1a44145db24be73634e36c39c564698