AllocStackInst wasn't invoking the parent copy constructor
Summary:
The "copy constructor" of an instruction is used for inlining,
ordinarily all it has to do is call the parent copy constructor, which
copies everything.

`AllocStackInst` is "special" because its operand is not heap allocated
(an admittedly questionable design choice, which however I would not
like to revisit now). So, invoking the parent copy constructor would
copy an operand pointing into a field of the source instruction.

So, `AllocStackInst`'s copy constructor attempts to perform the copy
itself without calling the parent copy constructor. In doing that, it
forgets to copy the type (as well as the location, etc).

Fix this by calling the parent copy constructor with no operands, and
manually append the single operand.

Reviewed By: avp

Differential Revision: D48932250

fbshipit-source-id: 8b9988f0504adeab882a2d6e72ed739e9a5f8b9e