static_h.h: add builtins for reading/writing native pointers
Summary:
`_sh_ptr_write_ptr(dest, offset, value)` and
`_sh_ptr_read_ptr(src, offset)`.

The offset is in units of `sizeof(void *)`, so assuming the base pointer
is correctly aligned (implied by its type in the function declaration),
the memory access is also guaranteed to be aligned.

Whether this is a good API remains to be seen. This is just a starting
point.

The implementation of the functions is interesting because it tries to
avoid UB concerns by using memcpy(). There is a long comment explaining
the details.

Reviewed By: avp

Differential Revision: D48932577

fbshipit-source-id: 990607507a0fe1e5fa752e9d07d9ceaf655a3c4c