Properly check method override types
Summary:
Method override typechecking was proceeding in the wrong direction.
The overriding method 'this' can be more specific than the super method,
so we have to handle `this` separately in the flow check.

Doing this correctly exposed a problem where we would then have to check
the super classes of the `this` argument prior to them being
initialized if we were to do the check during method type parsing.
So move the override check out into the checking visitor.

Reviewed By: tmikov

Differential Revision: D47959033

fbshipit-source-id: a4c77bdc2ea6b3120675ac9f78488411848448ec