Enable parsing of renders and component keywords in TypeAnnotation or Identifier contexts
Summary:
Today the following code fails with a parse error:
```
type T = (component: string) => string;
```
This happens because in a type function param position the identifier is optional. So we first parse as a TypeAnnotation and then "reparse" it to an identifier if a colon is found. This saves us doing a lookahead on all parameters. However for non simple annotation types such as ComponentType (e.g. `component(): void`) the initial TypeAnnotation parse will fail since a simple identifier does not match the expected format. This effectively makes these keywords reserved in this context, which makes rolling out new types much harder as it fails on existing code.

This change solves the issue by introducing a new function, `parseTypeAnnotationBeforeColonFlow`, which does a lookahead if a known conficiting ident is found to check for the colon (or question). If it is found it will return a `GenericTypeAnnotation` with the ident which can be unwrapped by `parseTypeAnnotationBeforeColonFlow`.

Reviewed By: avp

Differential Revision: D48710156

fbshipit-source-id: 1e0642b2bd726afc77e44a8c8cd3216179aaca63