Don't capture `this` directly in console HostFunctions
Summary:
Fix a bug in `CDPHandler`'s implementation of console logging
functions. Currently upon initialization, `CDPHandler` will overwrite
all of the console functions with a shim `HostFunction`. This
`HostFunction` captures the `this` of `CDPHandler::Impl`. It uses
`this` to emit the proper CDP event. In the following timeline-

1. CDPHandler is created
1. `console.log` is now a `HostFunction`
1. CDPHandler is destroyed
1. `console.log` is invoked

there will be an invalid memory access. `console.log` is never unbound
from the global object, so the `HostFunction` will get called and use
member variables of a now deallocated instance. We fix this by
capturing a `weak_ptr` to `this` instead in the installed `HostFunction`s

Reviewed By: mattbfb

Differential Revision: D49022186

fbshipit-source-id: 7e360438968e526d49b9d280d8d3f5e5c867bceb