Introduce SHModule and SHLocal
Summary:
## SHLocal
`sh::SHLocal` is an IR representation of a SH register. I decided to
call it a "local" instead of a "register" to avoid naming confusion with
sh::Register, but also because it is not actually a register.

It is a wrapper around an `sh::Register` with a type.

Instances of `sh::SHLocal` are uniqued based on the register and its
type. Thus, multiple `SHLocal` instances with the same register can
exist with different types.

## SHModule

SHModule is a new class that owns IR values necessary for SH lowering,
specifically `SHLocal`. It was introduced to avoid polluting the global
IR namespace with SH-specific types like registers, register classes,
etc.

The declaration of `sh::Register` had to be moved to IR to avoid
cyclical dependencies. However it is simple enough and only becomes
visible when `IR/SH/SHModule.h` is included.

There are no functional changes in the code.

Differential Revision: D49221050

fbshipit-source-id: a293062fff15dfb0d69d40634b853616fd09ff35