RegAlloc: replace allocated operands with SHLocal (#1126)
Summary:
Pull Request resolved: https://github.com/facebook/hermes/pull/1126

Replace operands referring to instructions with an instance of SHLocal
with the corresponding register and type. This breaks the usage
dependency between operands and instructions, which was logically
inconsistent after RegAlloc, since at that point the IR is no longer
SSA.

This now allows registers and types to be freely updated in individual
instructions.

PhiInst and AllocStackInst are removed.

This all happens after the lowering passes, since those passes still
need the dependency information left overt from SSA.

Differential Revision: D49221051

fbshipit-source-id: 588a656caa30dd16d4044ce0a1cbed4846b71113