Validate pause locations
Summary:
The test helpers contain a function to receive a
`Debugger.paused` notification and validate the
location where the pause occurred.

There are some tests that manually receive a pause
and manually the location, and others that receive
a pause and don't validate it was in the expected
location.

Use the helper the validates the location where
reasonable.

Reviewed By: fbmal7

Differential Revision: D49159587

fbshipit-source-id: e194a8da94931570b695bbdd519e526df7e67a3c