Test heap snapshot messages
Summary:
Update CDP tests to consume the heap snapshot notifications that were
previously untested.

A snapshot request generates a sequence of notifications followed by an
OK response. Ignoring the notifications in a test is possible because
the test setup buffers notifications and responses separately, allowing
the response to be tested (by `send(conn, req)`), while leaving the
preceding notifications unaccounted for.

A subsequent diff in this stack will combine notifications and
responses so they won't be accidentally ignored.

Reviewed By: dannysu

Differential Revision: D49188968

fbshipit-source-id: 291d19f4d442fe460344b2f3ff159fda7275e703