Test message message order
Summary:
Some tests claim that console and debugger notifications can
arrive in an arbitrary order, despite the `console.log` statement
and `debugger;` statement being strictly ordered.

To account for this, these tests receive multiple messages and
maintain flags to track what was received, instead of just expecting
notifications in the order that the code executes.

These messages should be ordered. e.g. in the test that runs
```
console.warn('string value', object1, object2);
debugger;
```
the console notification event should be sent before the
debugger paused event.

Adjust the tests to expect these notifications to be ordered.
This not only tests for proper ordering, but also simplifies the
test, as the state flags and receiving loops are no longer needed.

The one tricky bit to this diff is that the tests send and receive
messages to inspect remote object state. For the `expectProps`
function to receive the responses it expects, the other messages
in the incoming message queue must be processed first.

Thus, if some code produces multiple notifications, all of these
must be received before the test can start inspecting remote
objects with `expectProps`.

Otherwise, this diff simply unrolls the loops that were trying
to account for unpredictable message ordering.

Reviewed By: dannysu

Differential Revision: D49159562

fbshipit-source-id: a10d6ae3663d337821624748076f5907354aafe0