Fix heap snapshotting
Summary:
The CDP handler uses a custom stream that buffers data and emits it in
chunks that fit nicely into messages. The final chunk isn't emitted
until the stream is closed, emitting the final partially-filled chunk.

Chrome Dev Tools expects all chunks to be emitted before the heap
snapshot request is acknowledged with an "OK" response.

Make sure the stream is closed (i.e. all chunks are emitted) before
sending the "OK".

Although the tests appear to enforce the correct behavior, this is not
properly testable right now, as the test setup buffers responses and
notifications separately, making it impossible to test which arrived
first (the last chunk notification or the OK response). This will be
fixed later in this stack when messages are ordered.

Reviewed By: fbmal7

Differential Revision: D49159711

fbshipit-source-id: 164aa54f332fb813bf869711f5be314a61b18420