Order responses and notifications
Summary:
In tests, incoming notifications and responses are buffered separately.
This makes tests appear to verify things they actually don't. e.g. if a
test contains something like:
```
expectNotification<N>();
expectResponse<R>();
```
it looks like it's testing that we received notification of type `N`, followed
by response of type `R`. This order is not actually verified today.

If both the notification and response have arrived by the time this test
code runs, the test would also pass if rewritten to be:
```
expectResponse<R>();
expectNotification<N>();
```
because `expectResponse` just checks that the queue of incoming
responses contains a message of the the expected type (`R`), and
`expectNotification` checks that the queue of incoming
notifications contains a message of the expected type (`N`). If both
are already present and waiting the their respective queues, there's
no way for the test to verify which arrived first.

This diff combines the queues into a single queue for all incoming
messages, preserving the order in which they were received. This
allows tests to test the ordering of received messages. This also
makes it more difficult to accidentally ignore messages in tests
(e.g. if some code produces a sequence of responses and then
a notification, the test can't just test for the notification and
ignore all responses, because the notification won't be at the
front of the incoming message queue until the responses have
been processed).

Previous diffs in this stack have adjusted existing tests to
account for messages that were previously missed, and orderings
that didn't reflect the actual order messages were received.

Reviewed By: dannysu

Differential Revision: D49161512

fbshipit-source-id: 74bbe58ef4a494c9f5ae16b4d33f26c2d4f0e04a