Avoid recursion in JSError construction
Summary:
The `jsi::JSError` constructor currently makes further JSI calls to
populate the message and stack values. These JSI calls may throw, which
may then lead to the `JSError` constructor being recursively invoked.

This is especially a problem because our stack overflow detection
itself reports an exception on overflow, but producing that exception
requires retrieving those properties, which will then attempt to throw
another exception because the stack has already overflowed. This
continues until the native stack is exhausted and the program crashes.

To address this, use Hermes internal APIs to retrieve the value of
those properties. If this fail, we can simply suppress the exception
without constructing another `JSError`.

Reviewed By: avp

Differential Revision: D49211075