Add native stack checking to Runtime (#1102)
Summary:
Pull Request resolved: https://github.com/facebook/hermes/pull/1102

NOTE: This is a port of two diffs,
because the first one alone would cause errors.

Original Author: avp@meta.com
Original Git: 7a1a615c614100fc661cee4e016e51012a75e79a

Add a flag to control whether the true native stack checking is enabled.
This controls the VM behavior, defaulted to on but it'll error when
trying to compile for Windows.

Add a function to `Runtime` to check for native stack overflow.
The function is split:
* a fast inline check which does two simple comparisons
* a slow check that resets the stack bounds if the fast check fails

The fast inline check may erroneously return `false`,
but that will only happen if a new thread overlaps the old saved stack.

The amount of gap we're allowed at the end of the stack is configured
via `RuntimeConfig`.

This new mechanism will replace the `nativeCallFrameDepth_` mechanism,
which was in use until now but was conservative by design.
The RAII classes for using it are updated to use the true check,
if it is possible.

Some tests were checking for native stack overflow.
They need to be updated to grow the stack larger and actually cause an
overflow.

Original Reviewed By: neildhar

Original Revision: D41601035

=====

Original Revision: D47921011

Reviewed By: tmikov

Differential Revision: D47875957

fbshipit-source-id: bb6068081c4413073936eb5dbbf1f17ceebb1b3f