Improve IRGen of scoped for-loop
Summary:
A scoped for-loop can be improved when we know that one of these
combinations don't capture loop variables:
- The update expression
- Both the test and update expressions
- All of the test and update expressions and the body

For now we use a very conservative heuristic that checks if they contain
any function or class expressions and statements. Clearly in the future
we could check which variables are actually captured, though it feels
that optimizations like that don't belong in IRGen.

Most scoped for-loops will  have non-capturing test and update
expressions, which should mean a meaningful improvement of code
generation quality.

Reviewed By: avp

Differential Revision: D49361083

fbshipit-source-id: de82dd818492b25cc17802d8e1391999bb7498da