Implement emission of CheckedCastInst
Summary:
It works by either "positive" or "negative" checking, i.e. checking
whether the input is one of a set of invalid types, or whether it is one
of a set of valid types, depending on which set is smaller.

It doesn't check types of objects obviously, but works for primary
types.

The error message could be improved in the future to depend on the input
and output types.

The emitted C looks like this:
```C
if (!(_sh_ljs_is_string(locals.t0) || _sh_ljs_is_double(locals.t0)))
  _sh_throw_type_error_ascii(shr, "Checked cast failed");
```

Reviewed By: avp

Differential Revision: D49446107

fbshipit-source-id: 732e248da2cbee418c026444016299ac0deb3e1f