Resend scripts on each enable
Summary:
Any time the debugger is enabled, the client should be informed of the scripts that are already running.

Prior to this diff, scripts won't be sent if the previous session didn't shut down gracefully (i.e. scripts would only be sent on subsequent sessions if the previous session sent a `Debugger.disable`).

Reviewed By: dannysu

Differential Revision: D49374052

fbshipit-source-id: 4088bea3a2d519957d021cd8333665c8706836c6