Remove strict paramter in defineNameLengthAndPrototype
Summary:
This is a follow up to D48205935. The parameter `strictMode` was only
passed in order to correctly populate `caller` and `arguments`.
However, those are now *never* defined on the function directly. So, we
can now safely remove this paramter as it does nothing. This propogates
to the rest of the code that was plubming the `strict`ness of functions
through to this function, which is now unnecessary.

Reviewed By: neildhar

Differential Revision: D49239598

fbshipit-source-id: e7f7f263c5b3d5ae8a3b365b49c443b1bdb8094e