Lazily init NativeJSFunction props
Summary:
Support lazy initialization for `NativeJSFunction`. This is done by
creating a new table that contains the function information for the
length and name, which are the properties that need to be defined
lazily. This does come with a slight size increase for
`NativeJSFunction`s that have their properties materialized, as
their properties will overflow into prop storage.

Reviewed By: neildhar

Differential Revision: D48923517

fbshipit-source-id: 6dc2fca4756c7c4dab779bff1b6d14929c6a94c4