Cache console messages at all times
Summary:
The way Chrome browser & devtools work is that the browser's CDP layer
caches the console messages all the time. When devtools closes and
reconnects, the CDP layer will emit `consoleAPICalled` events to try to
restore the previous context. The restored context is bound by the size
of the cache, so Chrome doesn't restore everything.

This diff updates our message cache to behave the same. Every single
message is cached. Not just the messages before the first devtools
connection.

There are a couple other scenarios to be aware of:
1. For the reload case where the CDP layer and the runtime are reset,
Chrome browser does not persist the console messages. For example,
refreshing the webpage being inspected will not restore previous
console logs. Our CDPHandler will behave the same and no additional
persistence are implemented across the reload scenario.
1. Devtools has client-side persistence across reload of the page being
inspected (but not reload of the devtools). The client-side persistence
can be turned on in the settings via the "Preserve log" option.

Reviewed By: mattbfb

Differential Revision: D49655659

fbshipit-source-id: 84fb8388023227e851f8a54462db2f809a17256a