Use persistent value numbers with -Xdump-between-passes
Summary:
Enhance `Namer` to support "persistent mode", which requires it to keep
state per-function (normally the state is reset between functions).

Keep an optional "persistent IR namer" in the Context.

Modify IRPrinter to automatically use the persistent namer from the
context, if provided (but can also be forced not to).

Finally, initialize the persistent namer if `-Xdump-between-passes` is
set.

The result is that printed instruction and BB numbers are preserved from
pass to pass, making it much easier to see what changed.

No tests are affected, since tests don't use `-Xdump-between-passes`.

## Important Note About Memory Reuse

It is possible that an IR value can be freed and another value be
allocated in the same address. This would cause the persistent IR namer
to associate the same number with what are essentially two different
values from two different runs.

This is **not** a correctness issue. The assigned numbers are just
labels and don't actually matter; we never dereference the freed
pointers. But it might be confusing in some cases.

Later diffs improve on this.

Reviewed By: fbmal7

Differential Revision: D49674184

fbshipit-source-id: eadad740cf7677d9d953ef366b525e676c9c8bb8