Improve the persistent IR namer
Summary:
Remember the Value kind that was associated with a generated number. If
the kind changes, assign a new number. This can happen if an instruction
is freed and a new different one is allocated with the same address.

Before every pass, garbage collect values in the namer that were never
visited in the previous pass. This keeps the map size down and decreases
the chance that a new instruction would accidentally match the address
of an old one.

Reviewed By: fbmal7

Differential Revision: D49715995

fbshipit-source-id: ffa179a8ed19fc67f7fff4d656f09e0d83c01182