Cleanup DomTreeDFS a little
Summary:
Update the doc comments, remove unnecessary methods, rename some locals.

I was tempted to go a little further, but did not:
- Rename StackNode to StackEntry. The word "node" is used to refer to
two different in the same context. But such a rename would create
rebasing conflicts.
- Instead of a SmallVector of pointers to a RecyclingAllocator<> of
StackNode, simply use a std::deque<StackNode>. The main difference of
such an approach is that in very deep graphs, the deque will free the
memory allocated for children blocks when popping. I did not want to
risk the effect of this on large graphs.

Reviewed By: avp

Differential Revision: D49807013

fbshipit-source-id: 22b30bb3ead2a81ef1829f23222318c7d8a7600e