Convert 'as' to a new token
Summary:
Fixes the parsing of multiple TypeScript as expressions in a row.
Previously `x as any as number` was parsed as `x as (any as number)`,
instead of `(x as any) as number`.

This was because we customized the precedence calculation for `as`
in the function getPrecedenceExcept,
but this custom precedence was not saved in the stack.
Later when we had an as expression in the stack,
we were not getting the right precedence level for it when doing
`getPrecedence(stack.back().opKind)`.

Fix this by checking for `identifier` specially and converting
the `as` identifier to a new token kind, `as_operator`, before pushing
onto the stack.
This can be extended to implement `satisfies` as well.

Reviewed By: tmikov

Differential Revision: D49700751

fbshipit-source-id: 68d4fd7552cb90c3196d546c53989a5f29ba7add