Truncate legacy extension keys from default locale on iOS/Mac (#1141)
Summary:
Pull Request resolved: https://github.com/facebook/hermes/pull/1141

iOS can sometimes return a default locale identifier that contains
extensions in the legacy form. Ideally, we should convert these into
the new style and respect them, but that isn't trivial to do.

We could do also just use the iOS version of the locale without
converting it. But this is potentially confusing, since the locale
won't follow a format that we actually accept. So in something like:

```
Intl.Collator(Intl.Collator().resolvedOptions.locale)
```

the inner collator and outer collator will have inconsistent behaviour.

For now, truncate the extensions so we can at least parse the first
part of the identifier and honour it.

Closes #1140

Reviewed By: avp

Differential Revision: D49870526

fbshipit-source-id: 893c3e1d95009ab6b32b2e2c41ef84855cedf28a