Maintain ordering of loaded scripts
Summary:
We used to keep track of the loaded scripts in an `unordered_map`. This means that when a client connects and sends us a Debugger.enable message, the scriptParsed notifications that we send are not the same order as when those loaded scripts appeared.

This is a problem because DevTools' [DebuggerModel.ts](https://github.com/ChromeDevTools/devtools-frontend/blob/main/front_end/core/sdk/DebuggerModel.ts#L756) treats the scriptParsed that appear later on as the newest version of the script.

For the React Native Fast Refresh scenario where the same script gets modified and loaded multiple times, we want DevTools to treat the last loaded version as the newest version.

Fixing this issue by switching to use a vector so that the behavior is more like we buffered the loaded scripts and then send the scriptParsed notifications in the order that they would have been sent if the debugger client was already connected.

Reviewed By: fbmal7

Differential Revision: D50236422