Enable exceptions by default in CMake MSVC build
Summary:
CMake enables exceptions by default when using MSVC in order to
maintain compatibility with other compilers. However, we have a line
that removes this from the compiler flags, which means that targets by
default do not have exceptions enabled when compiling with MSVC.

However, this discrepancy is prone to mistakes. For instance,
`libhermes` is defined using `add_library` instead of
`add_hermes_library`. As a result, we have been building it with
exceptions disabled, which is incorrect.

Reviewed By: tmikov

Differential Revision: D48982162

fbshipit-source-id: 673c4572489af1a8d0a1d32299a7e07909597909