Support for renders variants
Summary:
Added support for `renders?` and `renders*`, along with the new layout for component syntax ASTs which sticks a TypeOperator in the `rendersType` field instead of a more general TypeAnnotation.

I updated the test cases to print render types as React.Node in the babel output and added new tests for pieterv to use when he goes to test prettier support.

Reviewed By: pieterv

Differential Revision: D50234180

fbshipit-source-id: 793602bb8e48b1aa6e8b78f247d82bdf0de63efb