Remove exceptions from synthTraceParser (#1159)
Summary:
Pull Request resolved: https://github.com/facebook/hermes/pull/1159

The synth trace parser has had a longstanding problem where it uses
exceptions to report errors, but also needs to depend on the Hermes
JSON parser, which means it needs to be compiled without RTTI. This
causes problems in executables that link against this library, because
it contains its own copy of type information for exception types. As a
result, `std::exception`s thrown across shared library boundaries may
not be catchable when one side of the boundary links against
synthTraceParser.

To avoid this, remove the use of exceptions in the synth trace parser
and replace them with fatals. The parser is only ever used to parse
generated traces, and the errors it reports are not recoverable in a
meaningful way, so exceptions are not particularly useful.

Reviewed By: tmikov

Differential Revision: D50294404

fbshipit-source-id: 5070cd2f48ffe5f834cff7565cc2cadd3619510a