Fold inspector into Hermes API libraries
Summary:
The inspector and our API libraries have a circular dependency.
Inspector code relies on JSI and the Hermes' debugger API via the
`hermesapi` static library, while Hermes' API libraries depend on the
inspector. This becomes a potential problem because it is very easy to
duplicate code and violate ODR. For example, a lean build of
`libhermes` may depend on the inspector, which in turn depends on the
non-lean `hermesapi`.

To avoid this, eliminate the inspector library, and directly include
its source files in Hermes' API libraries.

Reviewed By: dannysu

Differential Revision: D50251133

fbshipit-source-id: 84fdd22c4e0f67c7eb8a9e28f3d43128804e8efe