Replace hermesapi with libhermes (#1161)
Summary:

Delete the hermesapi library, in favour of always using libhermes. This
ensures that we are always testing libhermes itself, which is what
integrators actually use, and reduces the potential for confusion and
ODR violations introduced by `hermesapi`.

This should also avoid us creating further problems with exceptions and
RTTI of the kind that were addressed earlier in this stack, which may
have been obscured by linking statically.

Reviewed By: tmikov

Differential Revision: D50294403