Fix double -> int conversion in truncateToInt32
Summary:
ARM64 has different behavior when the double value can't fit into
int64_t (2^63-1 instead of -2^63 on x86-64), and 2^63-1 can't be
represented precisely in double, so it's converted to 2^63. The
result is that a double value 2^63 still go through the fast path
and eventually is casted to int32_t and -1 is returned, which is wrong.

The solution is to use smaller width integer (so every value can be
represented in double). And also use __builtin_constant_p() to
avoid UB caused by constant propagation.

Reviewed By: avp

Differential Revision: D50331023

fbshipit-source-id: f01c5fa7422ad3e30456904a6cbc4e3ca789baac