Use AddEdgesCallback to report WeakRefs
Summary:
Edges of WeakRefs for heap snapshot can be added in AddEdgesCallback
of HeapSnapshotMetadata, instead of EdgeAddingAcceptor. This is a
necessary step to completely remove WeakRefAcceptor.

Note that the order of edges may be different, but this is OK as Chrome
Dev Tool shows it in its own order anyway.

Reviewed By: neildhar

Differential Revision: D51223793

fbshipit-source-id: 90fb9c74536152f88c8884df0ee3a1fb63a9f20d