Reduce flakiness of WeakMap cycle test
Summary:
To try and reduce the flakiness of the WeakMap cycle test, increase the
number of iterations of the loop and allow larger heaps. This allows
the test to run with a higher steady state memory consumption, but
still ensures that no leaks occur.

This comes at the cost of a slower test.

Reviewed By: lavenzg

Differential Revision: D51485241

fbshipit-source-id: 4ecaa14fe8bb3d927772d2d4065d4356688edd8a