Introduce SynthTraceMode in RuntimeConfig
Summary:
For the next change to replace `HermesInternal.getInstrumentedStats`, we
need to be able to distinguish whether the runtime is doing synth
tracing and/or replaying the trace. Current `RuntimeConfig::TraceEnabled`
is boolean and can't express that the runtime is configured for synth
replay. Replace `RuntimeConfig::TraceEnabled` with `RuntimeConfig::
SynthTraceMode`, which is enum so that we can distinguish 4 types of
configuration.

Reviewed By: neildhar

Differential Revision: D51142856

fbshipit-source-id: 8e23fb201ca6d81af355fd46e73f377697a2d0dd