Use untraced host function call for HermesInternal.getInstrumentedStats
Summary:
At the start of a synth trace, replace non-deterministic
HermesInternal.getInstrumentedStats with HostFunctions that then call
the real methods without tracing. By intercepting these calls using a
HostFunction, the calls to this functions will be recorded as
HostFunction calls by the existing trace machinery, allowing us to
delete the MockedEnvironment mechanism in the VM. This simplifies synth
traces, and should mean that we no longer need to keep track of tracing
in the VM.

In the case of getInstrumentedStats(), which returns an Object type,
invisible to the synth trace during replay, the wrapping host function
intentionally reconstructs the object using jsi. This deliberate
recreation ensures that the synth trace records the object creation
process, providing insight into how the object was originally formed.

Reviewed By: neildhar

Differential Revision: D51475456

fbshipit-source-id: d5324e94fadd08b7bfc6661be9ad24591220c978