Update codegen to output type guards instead of `%checks`
Summary:
This diff updates the hermes-parser codegen to output type guards instead of `%checks` for the generated predicates file.

The current setup is in a weird state (we can't use hermes-parser to parse new syntax, because we are in hermes-parser itself), so I have to use flow-comments and disable prettier in the generated predicates file (prettier is on 2.8.8, which uses an old bundled-in flow-parser that doesn't support the new syntax). This will be a blocker for removing the remaining `%checks` or for adopting any new flow syntax in the hermes-parser codebase. At some point, we should just deal with the pain and move everything to prettier 3

Reviewed By: alexmckenley

Differential Revision: D51512414

fbshipit-source-id: 7f46d4cbb45559c23915ca221f764aea4af163f9