Fix Proxy defineProperty trap
Summary:
When a data property of a proxied object is assigned a new value, the defineProperty trap of Proxy is triggered with a descriptor with configurable: false, enumerable: false, and writable: false. However, according to the ECMAScript spec (see item 2.d.iii), the descriptor object must have no field other than value. This is problematic if the Proxy attempts to call Reflect.defineProperty() using the passed descriptor, as it unexpectedly "freezes" the property.

This is happening because `objectFromPropertyDescriptor` always construct an object with the fields set for writeable, enumerable, and configurable as `ComputedPropertyDescriptor` doesn't have notion of whether these fields exist or not.

Fixing this by changing to pass `DefinePropertyFlags` instead to `objectFromPropertyDescriptor` so that it can set the fields based on the existence of the fields.

Reviewed By: fbmal7

Differential Revision: D51327887

fbshipit-source-id: f77c4934b13fa1d4e5fda6b54c47beac92c9a56e