Remove asserts in objectFromPropertyDescriptor
Summary:
It turned out that these assertions hit with test262 tests. These test
cases are, for example, passing property descriptor with incomplete
attributes. For some tests, this doesn't matter and they are just
testing the behavior of trap functions. Few test cases are testing with
incorrect attribute set up and checking that the runtime is handling as
spec says. Essentially, `objectFromPropertyDescriptor` should create the
descriptor object as specified by the code, and Proxy should pass it to
the `defineProperty` trap function as is. Each of the case handling is
already well  implemented in the Proxy code.

Reviewed By: tmikov

Differential Revision: D51692787

fbshipit-source-id: 340b49007ca7a35eacab07be1c1e40a6e5b7ec7d