InstSimplify: fix handling of undefined in relational comparisons (#1202)
Summary:
Pull Request resolved: https://github.com/facebook/hermes/pull/1202

Relational comparisons (<= and >=) are simplified to a true/false when
it is known that the operands are the same value with a primitive type.
If the value is "undefined", it is converted to NaN, which never equals
itself.
The optimization was not correctly handling the case where the operand
could be "undefined" or another type - it was only checking for
"undefined".

Example:

    try {} catch {
      var v0 = "m";
    }
    return v0 <= v0;

The try/catch is necessary to make the optimizer believe `v0` is not a
constant.

Reported in https://github.com/facebook/hermes/issues/1200.

Reviewed By: neildhar

Differential Revision: D51555324

fbshipit-source-id: 87a4285166a9e11e96014feae45497d1a30d8307