Always write result last in the interpreter (#1203)
Summary:
If an instruction is the last user of one of its operands, the register
allocator may reuse one of its operands as the result register. This is
typically fine since most operands are only input parameters, and they
are always read before the output is written.

However, when the operand is both an input and an output, this requires
care to ensure that the operand register (whose live range ends with
the given instruction) is written before the result. This ensures that
even if the register allocator assigns them the same register, the
value held in that register after the instruction is the result.

Closes #1203

Reviewed By: avp

Differential Revision: D51597086

fbshipit-source-id: 14a4f27887d11cbe21bae5b27712b498b48c3f20